Refactor to packages per store type
